requirements.txt
altgraph==0.16.1
argh==0.26.2
auto-py-to-exe==2.6.5
bottle==0.12.17
bottle-websocket==0.2.9
cffi==1.12.3
Eel==0.10.4
future==0.17.1
gevent==1.4.0
gevent-websocket==0.10.1
greenlet==0.4.15
pathtools==0.1.2
pefile==2019.4.18
Pillow==6.1.0
pycparser==2.19
PyInstaller==3.5
pywin32-ctypes==0.2.0
PyYAML==5.1.2
schedule==0.6.0
tk==0.1.0
watchdog==0.9.0
whichcraft==0.6.0
